refactoring: PaperHeader폴더 이름 paperHeader로 변경 후 paper_components 폴더 안으로 위치 이동, constant.jsx파일명 constant.js로 수정
