+++ src/pages/paper/components/PaperHeader/EmojiChoice.style.js
