+++ src/pages/paper/components/PaperHeader/EmojiCountBtn.style.js
