+++ src/pages/paper/components/PaperHeader/PaperHeader.style.js
