+++ src/pages/paper/components/PaperHeader/TopEmojis.style.js
