+++ src/pages/paper/components/PaperHeader/constant.js
