+++ src/pages/paper/components/paperContents/PaperContents.style.js
