+++ src/pages/paper/components/paperContents/RollingPaperItem.style.js
