+++ src/pages/paper/components/paperContents/Skeleton.style.js
